+++ c14/example/src/main.rs
@@ -1,9 +1,9 @@
-mod c00_intro;
+mod s00_intro;
 
 fn main() {
     println!("Chapter 14");
     println!();
 
-    c00_intro::run();
+    s00_intro::run();
     println!();
 }
